+++ kotlin/reakt-native-toolkit/src/commonMain/kotlin/de/voize/reaktnativetoolkit/annotation/ExportTypescriptType.kt
@@ -1,0 +1,5 @@
+package de.voize.reaktnativetoolkit.annotation
+
+@Retention(AnnotationRetention.SOURCE)
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPEALIAS)
+annotation class ExportTypescriptType